04文献阅读/05肾病/13预测糖尿病酮症酸中毒相关急性肾损伤的危险因素/05复现/requirements.txt
pandas>=2.1
numpy>=1.26
pyyaml>=6.0
duckdb>=1.0.0
pyarrow>=15.0.0
scikit-learn>=1.4
xgboost>=2.0
matplotlib>=3.8
shap>=0.45
